Combine integration tests and unit tests into a single suite
