+++ .github/workflows/ci.yml
@@ -128,7 +128,6 @@ jobs:
           --health-retries 12
 
     env:
-      PARADEDB_INTEGRATION: "1"
       PARADEDB_TEST_DSN: postgresql://postgres:postgres@localhost:5432/postgres
       PGPASSWORD: postgres
       DATABASE_URL: postgresql://postgres:postgres@localhost:5432/postgres
@@ -154,11 +153,8 @@ jobs:
           uv venv --python "${{ matrix.python-version }}"
           uv sync --extra dev
 
-      - name: Run unit tests
-        run: uv run --with "${DJANGO_SPEC}" pytest -m "not integration" --cov=paradedb
-
-      - name: Run integration tests
-        run: uv run --with "${DJANGO_SPEC}" pytest -m integration --cov=paradedb --cov-append --cov-report=xml
+      - name: Run tests
+        run: uv run --with "${DJANGO_SPEC}" pytest --cov=paradedb --cov-report=xml
 
       - name: Upload to Codecov
         uses: codecov/codecov-action@v6
