+++ tests/test_diagnostics_integration.py
