+++ tests/test_error_handling.py
