+++ tests/test_facets_integration.py
