+++ tests/test_paradedb_queries.py
